tavily_tool/pyproject.toml
[project]
name = "tavily-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.8",
    "pydantic>=2.11.7",
    "python-decouple>=3.8",
    "tavily-python>=0.7.11",
]
